Rename new class and move function back
